+++ src/hyper_parameters.rs
@@ -1,3 +1,5 @@
+use std::fmt::Display;
+use num_traits::Num;
 use crate::distance::DistanceMetric;
 use crate::core_distances::NnAlgorithm;
 
@@ -5,6 +7,7 @@ use crate::core_distances::NnAlgorithm;
 const MIN_CLUSTER_SIZE_DEFAULT: usize = 5;
 const MAX_CLUSTER_SIZE_DEFAULT: usize = usize::MAX; // Set to a value that will never be triggered
 const ALLOW_SINGLE_CLUSTER_DEFAULT: bool = false;
+const EPSILON_DEFAULT: f64 = 0.0;
 const DISTANCE_METRIC_DEFAULT: DistanceMetric = DistanceMetric::Euclidean;
 const NN_ALGORITHM_DEFAULT: NnAlgorithm = NnAlgorithm::Auto;
 
@@ -12,32 +15,37 @@ const NN_ALGORITHM_DEFAULT: NnAlgorithm = NnAlgorithm::Auto;
 const MIN_CLUSTER_SIZE_MINIMUM: usize = 2;
 const MAX_CLUSTER_SIZE_MINIMUM: usize = 2;
 const MIN_SAMPLES_MINIMUM: usize = 1;
+const EPSILON_MINIMUM: f64 = 0.0;
 
 
 /// A wrapper around the various hyper parameters used in HDBSCAN clustering.
 /// Only use if you want to tune hyper parameters. Otherwise use `Hdbscan::default()` to 
 /// instantiate the model with default hyper parameters.
+#[derive(Debug, Clone, PartialEq)]
 pub struct HdbscanHyperParams {
     pub(crate) min_cluster_size: usize,
     pub(crate) max_cluster_size: usize,
     pub(crate) allow_single_cluster: bool,
     pub(crate) min_samples: usize,
+    pub(crate) epsilon: f64,
     pub(crate) dist_metric: DistanceMetric,
     pub(crate) nn_algo: NnAlgorithm,
 }
 
 /// Builder object to set custom hyper parameters.
+#[derive(Debug, Clone, PartialEq)]
 pub struct HyperParamBuilder {
     min_cluster_size: Option<usize>,
     max_cluster_size: Option<usize>,
     allow_single_cluster: Option<bool>,
     min_samples: Option<usize>,
+    epsilon: Option<f64>,
     dist_metric: Option<DistanceMetric>,
     nn_algo: Option<NnAlgorithm>,
 }
 
 impl HdbscanHyperParams {
-    pub(crate) fn default() -> Self {
+    pub(crate) fn default() -> HdbscanHyperParams {
         Self::builder().build()
     }
 
@@ -52,6 +60,7 @@ impl HdbscanHyperParams {
             max_cluster_size: None,
             allow_single_cluster: None,
             min_samples: None,
+            epsilon: None,
             dist_metric: None,
             nn_algo: None,
         }
@@ -72,8 +81,9 @@ impl HyperParamBuilder {
     /// # Returns
     /// * the hyper parameter configuration builder
     pub fn min_cluster_size(mut self, min_cluster_size: usize) -> HyperParamBuilder {
-        let valid_min_cluster_size = HyperParamBuilder::validate_input_left_bound(
-            min_cluster_size, MIN_CLUSTER_SIZE_MINIMUM, "min_cluster_size");
+        let valid_min_cluster_size = 
+            HyperParamBuilder::validate_input_left_bound(
+                min_cluster_size, MIN_CLUSTER_SIZE_MINIMUM, "min_cluster_size");
         self.min_cluster_size = Some(valid_min_cluster_size);
         self
     }
@@ -88,8 +98,9 @@ impl HyperParamBuilder {
     /// # Returns
     /// * the hyper parameter configuration builder
     pub fn max_cluster_size(mut self, max_cluster_size: usize) -> HyperParamBuilder {
-        let valid_max_cluster_size = HyperParamBuilder::validate_input_left_bound(
-            max_cluster_size, MAX_CLUSTER_SIZE_MINIMUM, "max_cluster_size");
+        let valid_max_cluster_size = 
+            HyperParamBuilder::validate_input_left_bound(
+                max_cluster_size, MAX_CLUSTER_SIZE_MINIMUM, "max_cluster_size");
         self.max_cluster_size = Some(valid_max_cluster_size);
         self
     }
@@ -118,9 +129,30 @@ impl HyperParamBuilder {
     /// # Returns
     /// * the hyper parameter configuration builder
     pub fn min_samples(mut self, min_samples: usize) -> HyperParamBuilder {
-        let valid_min_samples = HyperParamBuilder::validate_input_left_bound(
-            min_samples, MIN_SAMPLES_MINIMUM, "min_samples");
+        let valid_min_samples = 
+            HyperParamBuilder::validate_input_left_bound(
+                min_samples, MIN_SAMPLES_MINIMUM, "min_samples");
         self.min_samples = Some(valid_min_samples);
+        self
+    }
+
+    /// Sets epsilon, the distance threshold parameter. In HDBSCAN, each cluster has an epsilon
+    /// value, which is the distance threshold at which the cluster first appears, or the level
+    /// that it splits from its parent cluster. Setting this epsilon parameter creates a
+    /// distance threshold that a cluster must have existed at to be selected. If a cluster does 
+    /// not meet this threshold, it will be merged and the algorithm will search for a parent 
+    /// cluster that does meet the threshold. 
+    ///
+    /// # Parameters
+    /// * epsilon - the minimum distance threshold a cluster must have been in existence at
+    ///
+    /// # Returns
+    /// * the hyper parameter configuration builder
+    /// 
+    pub fn epsilon(mut self, epsilon: f64) -> HyperParamBuilder {
+        let valid_epsilon = 
+            HyperParamBuilder::validate_input_left_bound(epsilon, EPSILON_MINIMUM, "epsilon");
+        self.epsilon = Some(valid_epsilon);
         self
     }
 
@@ -165,12 +197,16 @@ impl HyperParamBuilder {
             max_cluster_size: self.max_cluster_size.unwrap_or(MAX_CLUSTER_SIZE_DEFAULT),
             allow_single_cluster: self.allow_single_cluster.unwrap_or(ALLOW_SINGLE_CLUSTER_DEFAULT),
             min_samples: self.min_samples.unwrap_or(min_cluster_size),
+            epsilon: self.epsilon.unwrap_or(EPSILON_DEFAULT),
             dist_metric: self.dist_metric.unwrap_or(DISTANCE_METRIC_DEFAULT),
             nn_algo: self.nn_algo.unwrap_or(NN_ALGORITHM_DEFAULT),
         }
     }
 
-    fn validate_input_left_bound(input_param: usize, left_bound: usize, param: &str) -> usize {
+    fn validate_input_left_bound<N>(input_param: N, left_bound: N, param: &str) -> N
+    where
+        N: Num + PartialOrd + Display
+    {
         if input_param < left_bound {
             println!(
                 "HDBSCAN_WARNING: {param} ({input_param}) cannot be lower \